Move validation to viewModel
